Fix third mgit function

+++ data/result.hpp
@@ -47,7 +47,7 @@ struct Result {
   double mgit_p;
   double mgit_midp;
   size_t mgit_successes;
-  size_t mgit_midp_successes;
+  double mgit_midp_successes;
   bool done;
   bool skippable;
   std::vector<double> permuted;
